+++ imports.rb
@@ -12,7 +12,6 @@ require_relative 'combat/beatdown'
 require_relative 'combat/boss_band'
 require_relative 'combat/boss_barkeep'
 require_relative 'combat/boss_bouncer'
-require_relative 'display/errors'
 require_relative 'display/menu_boss'
 require_relative 'display/menu_games'
 require_relative 'display/menu_main'
@@ -22,6 +21,7 @@ require_relative 'display/white_space'
 require_relative 'main/big_boss'
 require_relative 'main/escape_room'
 require_relative 'main/war_machine'
+require_relative 'messages/errors'
 require_relative 'messages/invoice'
 require_relative 'messages/narrative'
 require_relative 'messages/war_chatter'
